Feeling floaty

+++ hwTestSuite/tools.h
@@ -7,7 +7,7 @@
 
 Err makeFile(uint8_t* data, uint32_t size, char* fileName);
 uint16_t ads7846GetValue(uint8_t channel, Boolean referenceMode, Boolean mode8bit);
-double percentageOfTimeAs1(uint32_t address, uint8_t readSize, uint8_t bitNumber, uint32_t samples, uint32_t delay);
+float percentageOfTimeAs1(uint32_t address, uint8_t readSize, uint8_t bitNumber, uint32_t samples, uint32_t delay);
 var hexRamBrowser();
 var getTrapAddress();
 var manualLssa();
